Make the ubuntu14-x86_64 builder be triggered
It would be good to have an ubuntu machine in the set of
triggered builds for master, instead of just nightly.  It's
been running reliably and builds fast enough; we didn't add it
originally because there were some warnings that needed fixing.

Those warnings have now been fixed, so we can move it into the
triggered pool.