schedule more linux rc builds
Do daily builds with the linux rc builder for the branches,

  * master
  * openafs 1.8.x
  * openafs 1.6.x

for each,

  * the most recent linux rc
  * the daily linux snapshot made by the Ubuntu kernel team

This is 6 builds per day.

Do these in sequence instead of in parallel, since the script
seems to have a race condition bug that can fail when trying to
clone more than one kvm at a time.