Upgrade the linux-rc builder to ubuntu1804
Upgrade the nightly linux-rc builder to fix the issue with installing
the linux kernel ppa.

The virt-run command has been updated to support some new @ commands
to simplify the build command lines and some new options to set
the prefix of the ephemeral clone.

Comment out the smoke test until I'm confident the builds are working.