Get worker names from settings.ini
Create the workers list in a loop from the names and passwords listed in
the workers section of the settings.ini file. Since the worker names are
read from the ini file now, we now longer need to raise an exception
when the master.cfg and settings.ini file get out of sync.

A few of our workers have non-default options, such as max_builds and
notify_on_missing. Support this by listing just the ones we need in a
dictionary in the master.cfg file.