forcegerritbuild: Set event.change.url
Add new parameter to the ForceGerritBuild, gerriturl, that will be used
as a prefix for setting the event.change.url with the change ID and
if the patchset is given, the patchset.

There are 3 substitution fields:
   project
   changenumber
   patchsetnumber

The gerriturl should contain a string up to project name, e.g.
  gerriturl='https://gerrit.example.com:8080/#/c/%(changenumber)s/%(patchsetnumber)s'

Example generated url:
   https://gerrit.example.com:8080/#/c/4/6