roll-call: check against a list of known worker names
Buildbot returns old workers, which are expected to be offline.  Change
the roll-call script to report against a list of known workers and flag
an error if we do not get all of the known workers or if we get a new
worker not in our list of known workers.