remove master builds from old solaris 11 builder
This builder will be replaced by the new solaris 11.4 builder
soon for all of the branches. The 11.4 builder is not building
the stable branches yet, but is works for master branch builds,
so remove this old one from the gerrit master builds since
it is too old to build the master branch.