Retire the settings.ini file
Since we are using ansible, we can now drop the extra settings.ini file
to set the handful of host specific values and use host variable to set
those directly in the master.cfg file.

The three "mail to" addresses have been moved out of the repo and into
the local inventory file.

With this change it's harder to setup the master without ansible, since
the master.cfg is now a template, but the roles seem to work well now.