Fix incorrect conversion of patchset
The result from max can be an object and not just a number.  The value
returned by the max(patchsets...) is the actual patchset dictionary and
not just the patchset number.

Remove the incorrect conversion to str for the result of max.

Remove a redundant conversion  of an int object to an int.

Add the patchset number to the final msg for the buildbot log.