Move roles to a collection
Move the buildbot master and worker roles to the separate
github.com/openafs_contrib/buildbot repo, and import them as a
collection in ansible galaxy (openafs_contrib.buildbot).

Remove the old, pre-ansible instructions and scripts.

Remove non-master setup from the openafs_buildbot playbook and just
manage the master(s).

Add an inventory to manage the openafs master(s). Currently, we have an
old master and a new machine that is being staged to replace the
obsolete version.

Add a molecule scenario to locally test the openafs_buildbot (master)
playbook. The test scenario creates a test worker as well.