Add limit_gerrit_builds setting
Since the gerrit verification is all-or-none, sometimes we want to
temporarily remove builders from the gerrit pool. At the moment we have
two masters (a prod and one being staged to take over), so the list may
be diffent depending on the host.

To support this, add a master.cfg setting and ansible host variable
called 'limit_gerrit_builders' that has a list of which builders are
gerrit builders.