Fix buildbotURL on new master
I changed the buildbotURL in the inventory for the new master buildbot
while we were testing the gerrit integration, so the url's in the test
email notifications would not point to the old buildbot. Now that we are
close to cutting over, put the anticipated url back, and reset the
"delay before building" sleep time.