Move the git repo url to a site specific variable
To make it easier to setup buildbot master test systems, do not hardcode
the git repo url the master.cfg file, but instead make it a site
specific host variable.

For some reason some of the buildbots are adding a trailing slash
to the git url and the fetches are failing. Change the url from

    https://gerrit.openafs.org/openafs

to

    https://gerrit.openafs.org/openafs.git

to see if that helps.