Rename builders
The original OpenAFS buildbot setup did not distinguish between buildbot
builders (which are really build queues) and buildbot workers. This
makes it hard to see which builders are gerrit triggered and which are
nightly builders. It also makes it hard to see which branch was built on
the dashboards, and makes it difficult to have different build steps or
configure options depending on the branch or how the build was
scheduled.

Rename the builders so we can see which ones are gerrit builders and
which are nightly builders.  Also, add a branch type to the builder
suffix. Use the more general names 'dev', 'stable', and 'oldstable' for
the branches so we don't need to rename them when new major versions are
released.  Currently, these are associated with the git branch names
'master', 'openafs-stable-1_8_x', and 'openafs-stable-1_6_x'.