Park end-of-life fedora workers
We have a large number of end-of-life fedora workers, and it doesn't
make sense to use them as gerrit verifiers. But we need to keep a
builder config for them, otherwise they will fill the log with
connection exceptions.  Create a new category of builders called
"parked" for workers that are still online but we don't schedule gerrit
or nightly builds for any more.