Add WatchedWorker class to monitor unavailable workers
If a worker is unavailable, buildbot will not complete the report back
to gerrit.

Provide facility that will cancel a build job on a worker that is not
available.

Buildbot provides a python class, AbstractLatentWorker, for workers
that are created/destroyed on virtual hosts.  Part of this class is
support for handling worker timeouts.

Use the AbstractLatentWorker class to create a "WatchedWorker" class.

This class setups up a watchdog timer and will monitor for the actual
start of build steps. The watchdog timer is started when the build
starts and is turned off when any message is received from the worker.

To use this class:

add an import for watchedworker and use the watchedworker.WatchedWorker
instead of the worker.Worker class.

The length of time for the watchdog timer is controlled via the
missing_timeout keyword parameter (default set to 120 seconds).  The
length of the time needs to be sufficient to allow the worker to start
doing work, yet not too long as to hold up other builds.

When a build is cancelled, the final result status will be "exception"

Note:

It was determined that a build must be first started before it can be
cancelled so that the GerritStatusPush callback is invoked correctly.