Do not skip buildbot master install tasks
Remove the option to always skip the install tasks. This was a hold over
from the old server where we manually installed packages.  Just let
ansible run them and it will return no-changes if everything is already
up to date.

Since buildbot_master_install defaults to yes, just remove it from the
prod and test inventories.