add notify_on_missing for the workers
Set up the notify_on_missing facility to detect when a worker is not
available.  If a worker is "disconnected" for over 1 hour (the default
missing_timeout), one email will be sent to the "administrator" for that
worker.

Builds scheduled for these workers will remain in a "pending" state
until the worker reconnects.

Note if the buildbot master is restarted, the missing worker detection
is reset and emails will be resent for any missing workers.

Add the workers with their associated admins (admin emails obtained from
the current worker info)

When testing, the list of defined workers will be detected as missing.
To avoid sending emails to the admins while testing, define the ansible
variable "buildbot_test" in the molecule group vars.