Set WEB UI strings via ansible
The title for the top banner on the web page and the url link are
hardcoded strings.  In addition the settings.title is really being used
as the string for in gerrit summary message.

Set the WEB UI via ansible variables.
Add a new settings that contains the gerrit summary message title.

settings.title is the string that is displayed at the top of the
buildbot web page (and is limited in length by buildbot).

settings.summary_title is the string that is displayed in the build
summary message sent back to gerrit.