Fix shell cmd syntax error and make the step idempotent
The "Disable certificate checking" step in the default converge fails
with a syntax error in the git command.  In addition the step will
always register a change since it's using "shell".

Make the "Disable certificate checking" idempotent by using "git_config"
instead of "shell" for the step.  This also fixes the command syntax
error.