tests: Disable have_sudo in molecule instance
Be sure to match the settings for the actual buildbot master, which does
not have sudo enabled.

This also fixes an false test error with the latest buildbot role, since
the role no longer installed the buidlbotctl helper when sudo is
enabled. In that case, the role just uses systemctl to start and stop
the buildbot.

Add prepare playbooks for any virtual machine setup steps which require
root before running the converge, which runs our playbook to be tested.