Move templates to the templates directory
These started as files, but have been converted to jinja2 templates, so
move them to the templates directory and given them a .j2 file name
extention.

The buildbot master role will automatically remove the .j2 extension
when creating the target config file on the buildbot master.Move
templates to the templates directory

These started as files, but have been converted to jinja2 templates, so
move them to the templates directory and given them a .j2 file name
extention

The buildbot master role will automatically remove the .j2 extension
when creating the target config file on the buildbot master.