Templatize master.cfg
Reorganize the playbook and the master.cfg to separate the code from the data
so we can use molecule to effectively develop and test the buildbot master
configuration and code extensions.

Move the buildbot master.cfg from files to templates/master.cfg.j2 and make it
a proper template.  Move the builder and work names from the template to
inventory variables.

Move the python code to a new python package called afsbotcfg, which is now
built and installed by the playbook.  Provide a new set of buildbot factories
to make it easier to configure the unix builders.

Move the gerrit ssh setup from the buildbot_master role to the playbook.
Move the build script installation from the buildbot_master role to the
playbook.

Remove the passwords.ini file, which was left over from the pre-ansible
days. The worker passwords and the admin passwords are now in vault
encrypted inventory yaml files.

Reorganize the inventory variables.  A new set of dictionary variables are
defined in the inventory which are used to render the master.cfg on the
buildbot master.  These variables are overridden in the molecule test scenarios
to test with different builders.

The default molecule scenario creates the master instance only, with the
same configuration (except passwords) as the production instance.

The unix-builder scenario creates a master instances and 2 buildbot worker
instances, and configures the master for the two test worker instances.

Fully qualified module names to appease ansible-lint and fix the rest
of the ansible-lint warnings.

Freeze the python package versions and collection versions.

Install epel-release when preparing a test host.