Use a sentinel file target for python packages
Use the .packages sentinel file to represent the pip installation of the
required packages on the controller, instead of using 'activate' as a
prereq (which was created by python -m venv).

The pip installation can be skipped by specifying an empty PACKAGES
symbol on the command line or touching .packages in the project
directory.

While here, make a few makefile tweaks.