Add WatchedWorker class to monitor unavailable workers
If a worker is unavailable, buildbot will not complete the report back
to gerrit.

Provide facility that will cancel a build job on a worker that is not
responding.

Buildbot provides a python class, AbstractLatentWorker, for workers
that are created/destroyed on virtual hosts.  Part of this class is
support for handling worker timeouts.

Use the AbstractLatentWorker class to create a "WatchedWorker" class.

This class setups up a watchdog timer and will monitor for the progress
of the build. The watchdog timer is started when the build starts and is
reset when any message is received from the worker.

To use this class:

Add an import for watchedworker and use the watchedworker.WatchedWorker
instead of the worker.Worker class.  There are parameters for setting
the timeout and how to report a timed out worker back to gerrit.

The length of time for the watchdog timer is controlled via the
missing_timeout and stalled_timeout keyword parameters.  The initial
period is the missing_timeout value (default 15 minutes), with
subsequent time periods of stalled_timeout (default set to 5 minutes).
The initial timeout should handle cases where a worker could be just
starting up, but once a worker has timed out once, the shorter time
period is used to avoid holding up further builds.  The stalled_timeout
value should be long enough to allow a worker to generated some message
traffic within the time period.

When a build is cancelled, the final result status for the worker will
be "cancelled".  The summary reported back to gerrit will have detailed
information.

The stall_fail (True | False) parameter can be used cause the build
to be reported as a failure instead of cancelled.

Note:

It was determined that a build must be first started before it can be
cancelled so that the GerritStatusPush callback is invoked correctly. So
it isn't sufficient to just ping the worker to before a build starts.