Revert "Don't build if there are no code changes"
Due to a limitation with the currently installed version of gerrit on
gerrit.openafs.org, we cannot use a 'NO_CODE_CHANGE' filter at
this time.

If a patchset is being built and a "no code change" commit is pushed
before the build finishes, the vote from the build is not applied.

NOTE:
When tested with the latest version of gerrit, this limitation is
not present.

This reverts commit 4ed383a07f305ebed77bfbdcb31784bcfdc2a59c.

 Conflicts:
	src/afsbotcfg/gerrit.py  -- Just white space