Add compiler to builder names
The compiler is part of the "platform" for builds, so add it to the
builder names so the build status also shows the compilers and not just
the OS and arch.

Introduce a jinja2 macro to format the builder names from a set of
attributes in the master.yml file for each builder, instead of a
free-form name.  Some of the builder names needed to be changed for
this, but now the builder naming is consistent.