Remove molecule.yml templates
Modern versions of molecule do not support git ignored configuration
files, so remove the templatization of the molecule.yml files and just
use vagrant as the driver for testing the playbook.

Update the the package versions to current versions of Ansible,
molecule, etc.

Update the Ansible collection versions and import the new
openafs_devel_kernel role (which was split out from the
openafs_devel role).

Fix the bit rot in the unix-builders molecule scenario.