Rework build factories and steps
Refactor the factories and steps to be more consistent and remove
duplication.  Reduce the number of factories and create a single main
factory called UnixBuildFactory that can handle all the optional
variations; objdir builds, run unit tests, render docs.

Add a setup to build man pages and add the -q option to regen so we
build the man pages after configure.