Rename worker.enable to worker.inservice
Since 'enable' and 'disable' are overloaded, use 'inservice' to mean we
expect the worker to be online and working, even if it is temporarily
offline.

Reformat the workers yaml in the master configuration to make it easier
to read the list of workers. Use compact yaml here to list one worker
per line and format it in columns to make it more readable and easier to
keep in sorted order.