Split workers yaml into per worker files
Split these yml files into individual files to make it easier to add,
remove, and change workers.

The afsbotcfg_workers variable is now a list of worker names, instead of
a dict of worker names and properties.

With this change, workers can be "parked" by moving the workers/<name>
directory to a "parked" directory.