Fix nightly builder branch configuration
All of the nightly builders were defaulting to the master branch. Fix
the nightly schedulers by removing the change-filter (which we dont need
for timed builds), and specify a minimal 'codebases' dict in the nightly
schedulers.

Specify an empty string for the casebase name, since that is the default
name and we don't currently have multiple codebases.

Specify just the branch name in the codebase (e.g. 'master',
'openafs-stable-1_8_x') and use the default 'revision' and 'repository'
keys. The revision will be the most recent tip of the named branch, and
the respository is areadly defined in the build steps (although it
probably makes sense to move that to the codebases in the schedulers.)