afsbotcfg: Add DummyWorker check to summaryCB
In prepartion to add support to skip builders that have no workers
available, if a build was completed by a "dummy" worker, then the report
the build as skipped for verification purposes.

This allows the buildbot to continue to verify on the workers that are
available and prevent stalling verification if one worker goes offline.

The critical builders must be configured to have redundant workers and
no dummy worker, so we dont just verify everything if all the workers go
offline.