Checkout code from the canonical git repo on gerrit.openafs.org
Change the protocol used to checkout changes on the workers from "git"
to "https".  The former is checking out code from the git mirror that is
updated by gerrit when changes are pushed.

It was hoped fetching from the mirror would lessen the load on the
gerrit process and improve it's uptime, but the downside is changes are
updated asynchronously, so sometimes the changes are not propagated to
the mirror before the workers try to fetch them (it seems).

We do have delay step to try to avoid this issue, but that makes all
the builds slower and the delay is not reliable.

We believe the https checkouts should not impact gerrit, so try to
use that protocol instead.  We still have a delay and scatter in
place, as well as a counting lock to limit the number of concurrent
fetches.

The AIX 6 worker seems to be too old to use https, so continue to use
git protocol on that worker.