Add the GerritChangeSource gitBaseURL
Set the gitBaseURL in the GerritChangeSource so the "repository" build
property in gerrit builds matches the actual git url used to checkout
the code on those builds.

The "repository" build property is current not used, but it is
confusingly different than the actual git repo url used in the build
steps, and in the future we may want to use the build property in forced
builds to build changes before they are submitted to gerrit.