Add --output-sync make to parallel builders
Add the GNU Make --output-sync=target option to builders that have a
version of make that are new enough to support that option.  This should
not slow down the build times, but should make it easier to read build
errors.

Use the jobs option to specify when the --output-sync is used, and for
older machines that do not support it, just turn off parallel builds for
now.