Containerize the playbook execution with podman
Adopt podman to run ansible-playbook and other maintenance tasks. The
container images contain all the required dependencies needed to run the
playbook, so should be faster and more reliable to setup.  Maintainers
will need to install only podman and the required ssh config and keys to
connect to the buildbot machine.

Also include an image to run a playbook test on a local container.  This
is just a basic test of the playbook for now. Later, we can simulate
workers and the gerrit system for more complete testing.