Add testing containers
Add podman containers to test the afsbotcfg.yml ansible playbook.  The
test containers are:

* fake-buildbot-master: run a local test master instance
* fake-buildbot-worker: simulate the buildbot workers
* fake-gerrit: simulate the gerrit code review system

The fake-buildbot workers stub out the commands for testing build steps,
they do not actually build anything.

The containers run in pod so they can be managed as a group and share a
localhost interface.

Create a temporary ssh key pair for the ansible test and put the key
pair in a podman volume so the public key can be shared between the
containers.  Using a temporary key avoids the need for any developer
setup to run the tests.

Use podman exec to access the running master for debugging:

    podman exec -ti fake-buildbot-master /usr/bin/bash