Unvault the admin passwords in the password template
Defer the unvaulting of the admin passwords to the password.py template.

This is more consistent with the worker password handing, which are
unvaulted from the per worker vaulted password files, and will make it
possible to have an alternate password template for testing which does
not require vault.

Move the templating that generates the list of admin emails from the
master template to the password.py.j2 to consolidate the access to the
vault encrypted names.