Add reallyclean target to remove images and packages
Add a target to remove the podman images pulled after running the deploy
and test targets, and the afsbotcfg package created from the build
target.

The podman pod infra container image is also removed, unless it is in
use by another pod. (podman rmi will not remove the image when it is in
use by another pod.) In order to do this, we name the pod infra
container, and inspect that name to save the image name to the
.afsbotcfg-infra file.