test: Add afsbotcfg-fake-worker-alma10 container
Add a container that can actually do a build in order to test changes to
the unix build factory.

We fake out the `uname -r` to return a fixed linux kernel version and
setup a symlink to the kernel headers to simulate a real machine which
does not require special configure flags.