test: Add TEST_BUILDS flag for testing the unix build factory
Add TEST_BUILDS to create a different fake worker for the alma10
builders in order to test the unix builder. This is a quick and dirty
change for testing, a better way to do this is needed.

To use this,

    make check TEST_BUILDS=yes

and then use the "force" build of a nightly builder for alma10.  (The
fake gerrit builds do not work for this yet, since the fake gerrit
doesn't provide a usable git sha1 at this time.)